+++ history.md
@@ -1,11 +1,11 @@
-<!--mdast setext-->
+<!--remark setext-->
 
 <!--lint disable no-multiple-toplevel-headings-->
 
 1.0.0 / 2015-08-20
 ==================
 
-*   Update dependencies, dev-dependencies ([1dc3f51](https://github.com/wooorm/mdast-reference-links/commit/1dc3f51))
+*   Update dependencies, dev-dependencies ([1dc3f51](https://github.com/wooorm/remark-reference-links/commit/1dc3f51))
 
 0.1.0 / 2015-07-12
 ==================